pyproject.toml
[project]
name = "mindjunkies"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "authlib>=1.5.1",
    "cloudinary>=1.42.2",
    "crispy-tailwind>=1.0.3",
    "dj-database-url>=2.3.0",
    "django>=5.1.7",
    "django-allauth[socialaccount]>=65.4.1",
    "django-categories>=2.0.0",
    "django-crispy-forms>=2.3",
    "django-htmx>=1.23.0",
    "django-template-partials>=24.4",
    "markdown>=3.7",
    "pillow>=11.1.0",
    "psycopg2-binary>=2.9.10",
    "python-decouple>=3.8",
    "sslcommerz-lib>=1.0",
]

[dependency-groups]
dev = [
    "django-browser-reload>=1.18.0",
    "django-silk>=5.3.2",
    "django-tailwind-cli>=4.1.0",
    "freezegun>=1.5.1",
    "model-bakery>=1.20.4",
    "pre-commit>=4.1.0",
    "pytest>=8.3.5",
    "pytest-cov>=6.0.0",
    "pytest-django>=4.10.0",
    "pytest-mock>=3.14.0",
    "pytest-xdist>=3.6.1",
    "requests-mock>=1.12.1",
    "ruff>=0.9.10",
]
prod = [
    "gunicorn>=23.0.0",
    "whitenoise>=6.9.0",
]

[tool.ruff]
exclude = ["migrations", ".venv", ".git", ".vscode", ".idea", "__pycache__", ".pytest_cache", ".mypy_cache", ".coverage"]
line-length = 119

[tool.ruff.lint]
extend-select = [
    "E501",
    #    "UP", # pyupgrade
    "F", # Pyflakes
    "E", # pycodestyle-errors
    "W", # pycodestyle-warnings
    "N", # pep8-naming
    "I", # isort
]

[tool.pytest.ini_options]
DJANGO_SETTINGS_MODULE = "project.settings"
django_find_project = true
python_files = ["test_*.py", "*_tests.py", "tests.py"]
testpaths = ["tests", "mindjunkies"]
addopts = "--reuse-db --tb=short --capture=no --cov=mindjunkies --cov-report=term-missing --cov-report=xml --cov-config=.coveragerc"
